+++ lib/request/my_fatoorah_request.dart
@@ -41,15 +41,6 @@ class MyfatoorahRequest {
   /// This url must return the base response of my fatoorah without any edits
   final String executePaymentUrl;
 
-  /// this will controles what happen after payment done
-  ///
-  /// `AfterPaymentBehaviour.None` the default value , nothing will happen
-  ///
-  /// `AfterPaymentBehaviour.AfterCalbacksExecution` will pop after payment done and error or success callbacks finish
-  ///
-  /// `AfterPaymentBehaviour.BeforeCalbacksExecution` will pop after payment done and before error or success callbacks start
-  final AfterPaymentBehaviour afterPaymentBehaviour;
-
   final _testToken =
       "cxu2LdP0p0j5BGna0velN9DmzKJTrx3Ftc0ptV8FmvOgoDqvXivkxZ_oqbi_XM9k7jgl3SUriQyRE2uaLWdRumxDLKTn1iNglbQLrZyOkmkD6cjtpAsk1_ctrea_MeOQCMavsQEJ4EZHnP4HoRDOTVRGvYQueYZZvVjsaOLOubLkdovx6STu9imI1zf5OvuC9rB8p0PNIR90rQ0-ILLYbaDZBoQANGND10HdF7zM4qnYFF1wfZ_HgQipC5A7jdrzOoIoFBTCyMz4ZuPPPyXtb30IfNp47LucQKUfF1ySU7Wy_df0O73LVnyV8mpkzzonCJHSYPaum9HzbvY5pvCZxPYw39WGo8pOMPUgEugtaqepILwtGKbIJR3_T5Iimm_oyOoOJFOtTukb_-jGMTLMZWB3vpRI3C08itm7ealISVZb7M3OMPPXgcss9_gFvwYND0Q3zJRPmDASg5NxRlEDHWRnlwNKqcd6nW4JJddffaX8p-ezWB8qAlimoKTTBJCe5CnjT4vNjnWlJWscvk38VNIIslv4gYpC09OLWn4rDNeoUaGXi5kONdEQ0vQcRjENOPAavP7HXtW1-Vz83jMlU3lDOoZsdEKZReNYpvdFrGJ5c3aJB18eLiPX6mI4zxjHCZH25ixDCHzo-nmgs_VTrOL7Zz6K7w6fuu_eBK9P0BDr2fpS";
 
@@ -155,7 +146,6 @@ class MyfatoorahRequest {
     this.expiryDate,
     this.supplierCode,
     this.invoiceItems,
-    this.afterPaymentBehaviour = AfterPaymentBehaviour.None,
   });
   Map<String, dynamic> excutePaymentRequest(int paymentMethod) {
     var data = {