+++ lib/response/payment_response.dart
@@ -3,8 +3,9 @@ part of my_fatoorah;
 class PaymentResponse {
   final PaymentStatus status;
   final String paymentId;
+  final String url;
 
-  PaymentResponse(this.status, [this.paymentId]);
+  PaymentResponse(this.status, {this.url, this.paymentId});
   @override
   String toString() {
     return "Status: $status     PaymentId: $paymentId";